requirements.txt
blinker==1.9.0
cachelib==0.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
dropbox==12.0.2
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
geojson==3.2.0
geopandas==1.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
mapbox-vector-tile==2.1.0
MarkupSafe==3.0.2
mercantile==1.2.1
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
ply==3.11
protobuf==5.29.3
pyclipper==1.3.0.post6
pykml==0.2.0
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.17.0
stone==3.3.1
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
